refactor(ai-proxy): Улучшить AI provider manager и кэширование
- provider_manager.rs: обновлена логика конвертации tools для разных провайдеров
- cache.rs: улучшена обработка кэша для AI запросов

Подготовка к полной интеграции Function Calling на backend

+++ src-tauri/src/video_compiler/commands/ai_api_proxy/cache.rs
@@ -254,7 +254,7 @@ impl AICacheManager {
       total_entries: total_entries as u32,
       total_hits: total_hits as u32,
       expired_entries: expired_entries as u32,
-      max_entries: self.config.max_entries as u32,
+      max_entries: self.config.max_entries,
       enabled: self.config.enabled,
     })
   }
